requirements.txt
daphne
celery
python-decouple
Django
channels
django-cors-headers
requests
xlrd
aiohttp
lxml
xlsxwriter
redis
channels-redis
elasticsearch
psycopg2-binary
djangorestframework
djangorestframework-jwt
flower